skip verification if host or password is empty
