+++ pkg/detectors/jdbc/jdbc_test.go
@@ -2,14 +2,11 @@ package jdbc
 
 import (
 	"context"
-	"fmt"
 	"os"
-	"strings"
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/kylelemons/godebug/pretty"
-	logContext "github.com/trufflesecurity/trufflehog/v3/pkg/context"
 	"github.com/trufflesecurity/trufflehog/v3/pkg/detectors"
 	"github.com/trufflesecurity/trufflehog/v3/pkg/engine/ahocorasick"
 )
@@ -182,301 +179,6 @@ func TestJdbc_FromDataWithIgnorePattern(t *testing.T) {
 			}
 			if diff := pretty.Compare(got, tt.want); diff != "" {
 				t.Errorf("Jdbc.FromDataWithConfig() %s diff: (-got +want)\n%s", tt.name, diff)
-			}
-		})
-	}
-}
-
-func TestParseSqlServerMissingCredentialsBug1(t *testing.T) {
-	tests := []struct {
-		name        string
-		subname     string
-		shouldBeNil bool
-		reason      string
-	}{
-		{
-			name:        "no password - should return nil (nothing to verify)",
-			subname:     "//examplehost.net;databaseName=QRG1;sendStringParametersAsUnicode=false",
-			shouldBeNil: true,
-			reason:      "no password present",
-		},
-		{
-			name:        "no host - should return nil (invalid connection)",
-			subname:     "//;password=secret123",
-			shouldBeNil: true,
-			reason:      "no host present",
-		},
-		{
-			name:        "no host and no password - should return nil",
-			subname:     "//;databaseName=QRG1",
-			shouldBeNil: true,
-			reason:      "no host or password present",
-		},
-		{
-			name:        "csm-1584-example",
-			subname:     "//examplehost.net;databaseName=QRG1;sendStringParametersAsUnicode=false;loginTimeout=4;applicationName=pdal-cat-hierarchy-loader-v1;",
-			shouldBeNil: true,
-			reason:      "no password present",
-		},
-		{
-			name:        "valid with both host and password - should succeed",
-			subname:     "//examplehost.net;password=secret123",
-			shouldBeNil: false,
-		},
-		{
-			name:        "valid with host:port and password - should succeed",
-			subname:     "//examplehost.net:1433;password=secret123",
-			shouldBeNil: false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			ctx := logContext.AddLogger(context.Background())
-
-			j, err := parseSqlServer(ctx, tt.subname)
-
-			if tt.shouldBeNil {
-				if j != nil {
-					t.Errorf("parseSqlServer() expected nil (%s), but got connection: %v",
-						tt.reason, j)
-				}
-			} else {
-				if j == nil {
-					t.Errorf("parseSqlServer() returned nil, expected valid connection. err = %v", err)
-				}
-				if err != nil {
-					t.Errorf("parseSqlServer() unexpected error = %v", err)
-				}
-			}
-		})
-	}
-}
-
-// This test demonstrates the username is ignored when parsing the JDBC URL. Instead the default username "sa" is always used.
-func TestParseSqlServerUserIgnoredBug2(t *testing.T) {
-	tests := []struct {
-		name         string
-		subname      string // the part after "jdbc:sqlserver:"
-		wantUsername string
-	}{
-		{
-			name:         "user parameter specified",
-			subname:      "//localhost:1433;user=myuser;password=mypass",
-			wantUsername: "myuser",
-		},
-		{
-			name:         "user id parameter specified",
-			subname:      "//localhost:1433;user id=admin;password=secret",
-			wantUsername: "admin",
-		},
-		{
-			name:         "no user specified - should default to sa",
-			subname:      "//localhost:1433;password=mypass",
-			wantUsername: "sa",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			ctx := logContext.AddLogger(context.Background())
-
-			j, err := parseSqlServer(ctx, tt.subname)
-			if err != nil {
-				t.Fatalf("parseSqlServer() error = %v", err)
-			}
-
-			sqlServerConn := j.(*sqlServerJDBC)
-			expectedPrefix := fmt.Sprintf("sqlserver://%s:", tt.wantUsername)
-
-			if !strings.Contains(sqlServerConn.connStr, expectedPrefix) {
-				t.Errorf("Connection string does not contain expected username '%s'\nGot: %s\nExpected to contain: %s",
-					tt.wantUsername, sqlServerConn.connStr, expectedPrefix)
-			}
-		})
-	}
-}
-
-func TestParsePostgresMissingCredentials(t *testing.T) {
-	tests := []struct {
-		name        string
-		subname     string
-		shouldBeNil bool
-		reason      string
-	}{
-		{
-			name:        "no password - should return nil (nothing to verify)",
-			subname:     "//examplehost.net:5432/dbname?user=admin",
-			shouldBeNil: true,
-			reason:      "no password present",
-		},
-		{
-			name:        "no host - should return nil (invalid connection)",
-			subname:     "///dbname?user=admin&password=secret123",
-			shouldBeNil: true,
-			reason:      "no host present",
-		},
-		{
-			name:        "no host and no password - should return nil",
-			subname:     "///dbname",
-			shouldBeNil: true,
-			reason:      "no host or password present",
-		},
-		{
-			name:        "valid with host and password - should succeed",
-			subname:     "//examplehost.net:5432/dbname?user=admin&password=secret123",
-			shouldBeNil: false,
-		},
-		{
-			name:        "valid with localhost and password - should succeed",
-			subname:     "//localhost/dbname?user=postgres&password=secret123",
-			shouldBeNil: false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			ctx := logContext.AddLogger(context.Background())
-			j, err := parsePostgres(ctx, tt.subname)
-
-			if tt.shouldBeNil {
-				if j != nil {
-					t.Errorf("parsePostgres() expected nil (%s), got: %v", tt.reason, j)
-				}
-			} else {
-				if j == nil {
-					t.Errorf("parsePostgres() returned nil, expected valid connection. err = %v", err)
-				}
-				if err != nil {
-					t.Errorf("parsePostgres() unexpected error = %v", err)
-				}
-			}
-		})
-	}
-}
-
-func TestParsePostgresUsernameRecognition(t *testing.T) {
-	tests := []struct {
-		name         string
-		subname      string
-		wantUsername string
-	}{
-		{
-			name:         "user parameter specified",
-			subname:      "//localhost:5432/dbname?user=myuser&password=mypass",
-			wantUsername: "myuser",
-		},
-		{
-			name:         "user and password specified",
-			subname:      "//myuser:mypassword@localhost:5432/dbname",
-			wantUsername: "myuser",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			ctx := logContext.AddLogger(context.Background())
-			j, err := parsePostgres(ctx, tt.subname)
-			if err != nil {
-				t.Fatalf("parsePostgres() error = %v", err)
-			}
-
-			pgConn := j.(*postgresJDBC)
-			if pgConn.params["user"] != tt.wantUsername {
-				t.Errorf("expected username '%s', got '%s'", tt.wantUsername, pgConn.params["user"])
-			}
-		})
-	}
-}
-
-func TestParseMySQLMissingCredentials(t *testing.T) {
-	tests := []struct {
-		name        string
-		subname     string
-		shouldBeNil bool
-		reason      string
-	}{
-		{
-			name:        "no password - should return nil",
-			subname:     "//examplehost.net:3306/dbname?user=admin",
-			shouldBeNil: true,
-			reason:      "no password present",
-		},
-		{
-			name:        "no host - should return nil",
-			subname:     "///dbname?user=admin&password=secret123",
-			shouldBeNil: true,
-			reason:      "no host present",
-		},
-		{
-			name:        "no host and no password - should return nil",
-			subname:     "///dbname",
-			shouldBeNil: true,
-			reason:      "no host or password present",
-		},
-		{
-			name:        "valid with host and password - should succeed",
-			subname:     "//examplehost.net:3306/dbname?user=root&password=secret123",
-			shouldBeNil: false,
-		},
-		{
-			name:        "valid with localhost - should succeed",
-			subname:     "//localhost/dbname?user=root&password=secret123",
-			shouldBeNil: false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			ctx := logContext.AddLogger(context.Background())
-			j, err := parseMySQL(ctx, tt.subname)
-
-			if tt.shouldBeNil {
-				if j != nil {
-					t.Errorf("parseMySQL() expected nil (%s), got: %v", tt.reason, j)
-				}
-			} else {
-				if j == nil {
-					t.Errorf("parseMySQL() returned nil, expected valid connection. err = %v", err)
-				}
-				if err != nil {
-					t.Errorf("parseMySQL() unexpected error = %v", err)
-				}
-			}
-		})
-	}
-}
-
-func TestParseMySQLUsernameRecognition(t *testing.T) {
-	tests := []struct {
-		name         string
-		subname      string
-		wantUsername string
-	}{
-		{
-			name:         "user parameter specified",
-			subname:      "//localhost:3306/dbname?user=myuser&password=mypass",
-			wantUsername: "myuser",
-		},
-		{
-			name:         "no user specified - default root",
-			subname:      "//localhost:3306/dbname?password=mypass",
-			wantUsername: "root",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			ctx := logContext.AddLogger(context.Background())
-			j, err := parseMySQL(ctx, tt.subname)
-			if err != nil {
-				t.Fatalf("parseMySQL() error = %v", err)
-			}
-
-			mysqlConn := j.(*mysqlJDBC)
-			if !strings.Contains(mysqlConn.userPass, tt.wantUsername) {
-				t.Errorf("Connection string does not contain expected username '%s'\nGot: %s\nExpected: %s",
-					tt.wantUsername, mysqlConn.userPass, tt.wantUsername)
 			}
 		})
 	}
